pyproject.toml
[project]
name = "gitops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "gunicorn>=21.2.0",
]
